Add more detailed info about units in review
This patch adds information about percentage of planned
and unplanned tasks in "in review" state.
Such tasks should be consiedered as "not done" of course
so they aren't included in "achieved" stats but it may
be useful to have info about how many of units are
in review, so may be consiedered as "almost done" in most
cases.

Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>