Add support for setting the PM_SCORE field
This patch is adding support for setting the PM_SCORE field for the
cards in the Backlog. For cards with the BUGZILLA field set, we will
fetch the PM_SCORE from that URL and add it to the Trello card, this
will help us prioritizing work during the planning.

This patch includes two new subcommands:

* report: Which does what this tool was suppose to do before, generate a
report about the Sprint

* pm-score: (see commit message)

Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>